+++ src/python/molpher/core/operations/callbacks.py
@@ -17,4 +17,4 @@ Just holds a reference to `TraverseCallback`
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from molpher.swig_wrappers.core import TraverseCallback
+from molpher.swig_wrappers.core import TraverseCallback, SortMorphsCallback